+++ packages/nuxt/src/mod.ts
@@ -1,0 +1,2 @@
+export { default } from "./module.ts";
+export type { ContextDataFactory, ModuleOptions } from "./module.ts";